requirements.txt
streamlit 
langchain-groq 
langchain-google-genai 
langchain-core
faiss-cpu
langchain-tavily
duckduckgo-search
pypdf
langchain-community
langchain-text-splitters
python-dotenv